requirements.txt
pydantic>=2.7,<3
pydantic-settings>=2.3,<3
fastapi>=0.111,<0.115
uvicorn[standard]>=0.30,<0.33
python-docx>=1.1,<1.2
openai>=1.52,<2
requests>=2.32,<3
sqlalchemy>=2.0,<3
greenlet>=3.0,<4
asyncpg>=0.29,<1
httpx>=0.27,<1
supabase>=2.0,<3
python-multipart>=0.0.6,<1
